style: :art: divide SafeValue.cs into 5 files
