+++ logic/Preparation/Utility/SafeValue/SafeValueDouble.cs
@@ -1,0 +1,8 @@
+﻿using System;
+using System.Threading;
+
+namespace Preparation.Utility
+{
+    //其对应属性不应当有set访问器，避免不安全的=赋值
+
+}
